refactor: move `host` to client dir
I can't decide where this belongs 😅 this feels better for now
